♻️ F13B: moved database api calls in a proper file
