+++ tsconfig.json
@@ -18,7 +18,8 @@
     "./src/**/*.tsx",
     "src/index.tsx",
     "src/components/pages/order/OrderPage.tsx",
-    "src/components/reusable/TextInput.tsx"
+    "src/components/reusable/TextInput.tsx",
+    "src/api/users.js"
   ],
   "exclude": ["node_modules"]
 }
